Write codebase documentation. Restructure the repo so that the doc makes sense. (#54)
