+++ src/mlagility/__init__.py
@@ -1,4 +1,4 @@
 from mlagility.version import __version__
 
-from .api import benchit
-from .cli import main as benchitcli
+from .api.api import benchit
+from .cli.cli import main as benchitcli
