+++ src/mlagility/analysis/util.py
@@ -9,7 +9,7 @@ import groqflow.justgroqit.export as export
 import groqflow.justgroqit.stage as stage
 from groqflow.common import printing
 import groqflow.common.build as build
-import mlagility.filesystem as filesystem
+import mlagility.common.filesystem as filesystem
 
 
 @dataclass
